% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaesan-package.R
\docType{package}
\name{jaesan-package}
\alias{jaesan}
\alias{jaesan-package}
\title{jaesan: Data Package about Property of People who Work(ed) for Korea}
\description{
Package provides data about property of 
            people who work or worked for Korea. 
            Preprocessed data is from  newstapa's jaesan site <https://jaesan.newstapa.org/>.
            Spacial thanks to newstapa <https://newstapa.org/>.
            Data License is cc by-sa 4.0 International *Copyright newstapa*.
}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (0000-0001-6474-2570)

}
\keyword{internal}
