% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{query name of peaple who work(ed) for Korea Government}
\usage{
query(name, msg = T)
}
\arguments{
\item{name}{query string for search name or title in gov.}

\item{msg}{announce root url. Defualt is TRUE.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
This function use newstapa site \url{https://jaesan.newstapa.org/}
}
