% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\docType{data}
\name{recode}
\alias{recode}
\alias{recode2011}
\alias{recode2012}
\alias{recode2013}
\alias{recode2014}
\alias{recode2015}
\alias{recode2016}
\alias{recode2017}
\alias{recode2018}
\title{jaesan data detail called recode.}
\format{tibble with columns}
\source{
https://jaesan.newstapa.org/
}
\usage{
recode2011

recode2012

recode2013

recode2014

recode2015

recode2016

recode2017

recode2018
}
\description{
Property of People who Work(ed) for Korea
}
\examples{
  recode2011
  recode2012
  recode2013
  recode2014
  recode2015
  recode2016
  recode2017
  recode2018
}
\keyword{datasets}
