% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search}
\alias{search}
\title{search name of peaple who worked for Korea}
\usage{
search(name_query)
}
\arguments{
\item{name_query}{query string for search name.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
This function use newstapa site \url{https://jaesan.newstapa.org/}
}
